claudecode-agent-sdk-tutorial/pyproject.toml
[project]
name = "cc-agent-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.3",
    "e2b>=2.3.2",
    "e2b-code-interpreter>=2.2.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]

[tool.uv]
managed = true
